definitions of some parameters changed

+++ HasKAL/src/HasKAL/DataBaseUtils/FrameFull/Data.hs
@@ -7,7 +7,7 @@ data CDFParam = CDFParam { cdf'samplingFrequency :: Double
                          , cdf'cutoffFrequencyLow :: Double
                          , cdf'cutoffFrequencyHigh :: Double
                          , cdf'blockSize :: Int
-                         , cdf'fftSize :: Int
-                         , cdf'chunkSize :: Int
+                         , cdf'fftSize :: Double
+                         , cdf'chunkSize :: Double
                          }
